refactor(renderer): extract chat model/thinking + header-alert derivations (Round E blade 1)
Move the model/thinking selection cluster (chatModelChoices, active/new-chat
model+label, thinking-variant lists, sticky-pick validation, both pending
new-chat states) and the two chat-header alert memos out of app-shell.tsx into
a new pure-derivation hook use-shell-chat-model.ts. openSettingsSection is
injected so chatConnectionAlert keeps wrapping the derived click target with
its identical (exhaustive-deps-excluded) dep array. Every memo keeps its exact
dependency array so chatModelChoices / activeThinkingLevels / newChatThinkingLevels
retain referential stability. Zero behavior change.

app-shell.tsx 1680 -> 1585. Re-pinned: added use-shell-chat-model.ts to the
renderer-shell-source combined allowlist and to the composer-new-chat model
picker contract's subset reads (pendingNewChatModel / validPendingNewChatModel /
newChatThinkingLevel declarations moved into the hook).
